+++ src/app/payments/index.ts
@@ -2,3 +2,4 @@ export * from "./get-protocol-fee"
 export * from "./send-lightning"
 export * from "./send-intraledger"
 export * from "./update-pending-payments"
+export * from "./reimburse-fee"
