+++ src/app/wallets/index.ts
@@ -10,8 +10,6 @@ export * from "./get-transaction-by-id"
 export * from "./get-transactions-by-addresses"
 export * from "./get-transactions-by-hash"
 export * from "./get-transactions-for-wallet"
-export * from "./reimburse-failed-usd"
-export * from "./reimburse-fee"
 export * from "./send-on-chain"
 export * from "./register-broadcasted-payout-txn"
 export * from "./settle-payout-txn"
